I believe I have all the files that I wanted updated here. I made a menu with basic buttons that change when hovered over or clicked. I'm still trying to fix my classpath but with how much other stuff I have to do that part is for another time. I will also have to look at why sometimes I can't move my character around.

+++ myPlatformer/src/entities/Entity.java
@@ -23,7 +23,7 @@ public abstract class Entity { //a class we can't make an object on
 		g.drawRect((int)hitBox.x, (int)hitBox.y, (int)hitBox.width, (int)hitBox.height);
 	}
 	
-	protected void initHitBox(float x, float y, float width, float height) { //sets the hit box
+	protected void initHitBox(float x, float y, int width, int height) { //sets the hit box
 		hitBox = new Rectangle2D.Float((int)x, (int)y, width, height);
 	}
